reorganize and remove unnecessary tests
