+++ featuretools/tests/entityset_tests/test_dask_es.py
@@ -4,70 +4,6 @@ import pytest
 
 import featuretools as ft
 from featuretools.entityset import EntitySet, Relationship
-from featuretools.primitives import (
-    get_aggregation_primitives,
-    get_transform_primitives
-)
-
-UNSUPPORTED = [p.name for p in get_transform_primitives().values() if not p.dask_compatible]
-UNSUPPORTED += [p.name for p in get_aggregation_primitives().values() if not p.dask_compatible]
-
-
-def test_transform(es, dask_es):
-    primitives = ft.list_primitives()
-    trans_list = primitives[primitives['type'] == 'transform']['name'].tolist()
-    trans_primitives = [prim for prim in trans_list if prim not in UNSUPPORTED]
-    agg_primitives = []
-
-    assert es == dask_es
-    # Run DFS using each entity as a target and confirm results match
-    for entity in es.entities:
-        fm, _ = ft.dfs(entityset=es,
-                       target_entity=entity.id,
-                       trans_primitives=trans_primitives,
-                       agg_primitives=agg_primitives,
-                       cutoff_time=pd.Timestamp("2019-01-05 04:00"),
-                       max_depth=2,
-                       max_features=100)
-
-        dask_fm, _ = ft.dfs(entityset=dask_es,
-                            target_entity=entity.id,
-                            trans_primitives=trans_primitives,
-                            agg_primitives=agg_primitives,
-                            cutoff_time=pd.Timestamp("2019-01-05 04:00"),
-                            max_depth=2,
-                            max_features=100)
-        # Use the same columns and make sure both indexes are sorted the same
-        dask_computed_fm = dask_fm.compute().set_index(entity.index).loc[fm.index][fm.columns]
-        pd.testing.assert_frame_equal(fm, dask_computed_fm)
-
-
-def test_aggregation(es, dask_es):
-    primitives = ft.list_primitives()
-    trans_primitives = []
-    agg_list = primitives[primitives['type'] == 'aggregation']['name'].tolist()
-    agg_primitives = [prim for prim in agg_list if prim not in UNSUPPORTED]
-
-    assert es == dask_es
-
-    # Run DFS using each entity as a target and confirm results match
-    for entity in es.entities:
-        fm, _ = ft.dfs(entityset=es,
-                       target_entity=entity.id,
-                       trans_primitives=trans_primitives,
-                       agg_primitives=agg_primitives,
-                       cutoff_time=pd.Timestamp("2019-01-05 04:00"),
-                       max_depth=2)
-
-        dask_fm, _ = ft.dfs(entityset=dask_es,
-                            target_entity=entity.id,
-                            trans_primitives=trans_primitives,
-                            agg_primitives=agg_primitives,
-                            cutoff_time=pd.Timestamp("2019-01-05 04:00"),
-                            max_depth=2)
-        # Use the same columns and make sure both indexes are sorted the same
-        dask_computed_fm = dask_fm.compute().set_index(entity.index).loc[fm.index][fm.columns]
-        pd.testing.assert_frame_equal(fm, dask_computed_fm, check_dtype=False)
 
 
 def test_create_entity_from_dask_df(es):
@@ -478,31 +414,6 @@ def test_single_table_dask_entityset_dates_not_sorted():
     pd.testing.assert_frame_equal(fm, dask_fm.compute().set_index('id').loc[fm.index])
 
 
-# def test_training_window_parameter(mock_customer_es, mock_customer_dask_es):
-#     entity = "customers"
-#     cutoff_times = pd.DataFrame()
-#     cutoff_times['customer_id'] = [1, 2, 3, 1]
-#     cutoff_times['time'] = pd.to_datetime(['2014-1-1 04:00',
-#                                            '2014-1-1 05:00',
-#                                            '2014-1-1 06:00',
-#                                            '2014-1-1 08:00'])
-#     cutoff_times['label'] = [True, True, False, True]
-
-#     cutoff_times_dask = dd.from_pandas(cutoff_times, npartitions=mock_customer_dask_es[entity].df.npartitions)
-
-#     dask_fm, _ = ft.dfs(entityset=mock_customer_dask_es,
-#                         target_entity=entity,
-#                         cutoff_time=cutoff_times_dask,
-#                         training_window="2 hour")
-
-#     fm, _ = ft.dfs(entityset=mock_customer_es,
-#                    target_entity=entity,
-#                    cutoff_time=cutoff_times,
-#                    training_window="2 hour")
-
-#     pd.testing.assert_frame_equal(fm, dask_fm.compute().set_index('customer_id'))
-
-
 def test_secondary_time_index():
     log_df = pd.DataFrame()
     log_df['id'] = [0, 1, 2, 3]
@@ -585,82 +496,3 @@ def test_secondary_time_index():
 
     # Make sure both matrixes are sorted the same
     pd.testing.assert_frame_equal(fm.sort_values('delay'), dask_fm.compute().set_index('id').sort_values('delay'))
-
-
-def test_build_es_from_scratch_and_run_dfs():
-    es = ft.demo.load_mock_customer(return_entityset=True)
-    data = ft.demo.load_mock_customer()
-
-    transactions_df = data["transactions"].merge(data["sessions"]).merge(data["customers"])
-    transactions_dd = dd.from_pandas(transactions_df, npartitions=4)
-    products_dd = dd.from_pandas(data["products"], npartitions=4)
-    dask_es = EntitySet(id="transactions")
-
-    transactions_vtypes = {
-        "transaction_id": ft.variable_types.Id,
-        "session_id": ft.variable_types.Id,
-        "transaction_time": ft.variable_types.DatetimeTimeIndex,
-        "product_id": ft.variable_types.Id,
-        "amount": ft.variable_types.Numeric,
-        "customer_id": ft.variable_types.Id,
-        "device": ft.variable_types.Categorical,
-        "session_start": ft.variable_types.Datetime,
-        "zip_code": ft.variable_types.ZIPCode,
-        "join_date": ft.variable_types.Datetime,
-        "date_of_birth": ft.variable_types.Datetime
-    }
-    dask_es.entity_from_dataframe(entity_id="transactions",
-                                  dataframe=transactions_dd,
-                                  index="transaction_id",
-                                  time_index="transaction_time",
-                                  variable_types=transactions_vtypes)
-
-    products_vtypes = {
-        "product_id": ft.variable_types.Id,
-        "brand": ft.variable_types.Categorical
-    }
-    dask_es.entity_from_dataframe(entity_id="products",
-                                  dataframe=products_dd,
-                                  index="product_id",
-                                  variable_types=products_vtypes)
-
-    new_rel = Relationship(dask_es["products"]["product_id"],
-                           dask_es["transactions"]["product_id"])
-
-    dask_es = dask_es.add_relationship(new_rel)
-
-    dask_es = dask_es.normalize_entity(base_entity_id="transactions",
-                                       new_entity_id="sessions",
-                                       index="session_id",
-                                       make_time_index="session_start",
-                                       additional_variables=["device",
-                                                             "customer_id",
-                                                             "zip_code",
-                                                             "session_start",
-                                                             "join_date",
-                                                             "date_of_birth"])
-
-    dask_es = dask_es.normalize_entity(base_entity_id="sessions",
-                                       new_entity_id="customers",
-                                       index="customer_id",
-                                       make_time_index="join_date",
-                                       additional_variables=["zip_code", "join_date", "date_of_birth"])
-
-    trans_primitives = ['absolute', 'is_weekend', 'year', 'day', 'num_characters', 'num_words']
-    agg_primitives = ['count', 'max', 'sum']
-
-    fm, _ = ft.dfs(entityset=es,
-                   target_entity="customers",
-                   trans_primitives=trans_primitives,
-                   agg_primitives=agg_primitives,
-                   max_depth=2)
-
-    dask_fm, _ = ft.dfs(entityset=dask_es,
-                        target_entity="customers",
-                        trans_primitives=trans_primitives,
-                        agg_primitives=agg_primitives,
-                        max_depth=2)
-
-    # Use the same columns and make sure both have same index sorting
-    _dask_fm = dask_fm.compute().set_index('customer_id').reindex(fm.index)
-    pd.testing.assert_frame_equal(fm, _dask_fm[fm.columns], check_dtype=False)
